project/pipeline/minerwa/pyproject.toml
[tool.poetry]
name = "minerwa"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
orjson = "3.8.4"
"oslo.config" = "9.0.0"
pynsq = "0.9.1"
nats-py = "^2.2.0"
pycapnp = "^1.3.0"
pyzmq = "^25.1.1"

[tool.poetry.dev-dependencies]

[tool.poetry.plugins."minerwa.datasources"]
nsq = "minerwa.datasources.nsq:NSQDatasource"
zmq = "minerwa.datasources.zmq:ZMQDatasource"

[tool.poetry.plugins."minerwa.ingestor_processors"]
vflow = "minerwa.ingestors.processors.vflow:VFlowProcessor"
nprobe = "minerwa.ingestors.processors.nprobe:NProbeProcessor"

[tool.poetry.plugins."minerwa.detectors"]
ai = "minerwa.detectors.ai:AIDetector"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
